City now has a land value for each tile, and this can be displayed. Right now it's 128 everywhere because I don't yet have a calculation for it. CLOC: 14143

+++ src/render.cpp
@@ -558,7 +558,7 @@ void endRectsGroup(DrawRectsGroup *group)
 	endCurrentSubGroup(group);
 }
 
-void drawGrid(RenderBuffer *buffer, Rect2 bounds, s8 shaderID, s32 gridW, s32 gridH, u8 *grid, u8 paletteSize, V4 *palette)
+void drawGrid(RenderBuffer *buffer, Rect2 bounds, s8 shaderID, s32 gridW, s32 gridH, u8 *grid, u16 paletteSize, V4 *palette)
 {
 	DEBUG_FUNCTION_T(DCDT_Renderer);
