+++ src/render.h
@@ -85,7 +85,7 @@ struct RenderItem_DrawRects_Item
 struct RenderItem_DrawGrid
 {
 	Rect2 bounds; // Rendering position of the grid as a whole
-	u8 paletteSize;
+	u16 paletteSize;
 	s32 gridW, gridH; // Dimensions of the array
 };
 
@@ -242,7 +242,7 @@ void endRectsGroup(DrawRectsGroup *group);
 DrawRectsSubGroup beginRectsSubGroup(DrawRectsGroup *group);
 void endCurrentSubGroup(DrawRectsGroup *group);
 
-void drawGrid(RenderBuffer *buffer, Rect2 bounds, s8 shaderID, s32 gridW, s32 gridH, u8 *grid, u8 paletteSize, V4 *palette);
+void drawGrid(RenderBuffer *buffer, Rect2 bounds, s8 shaderID, s32 gridW, s32 gridH, u8 *grid, u16 paletteSize, V4 *palette);
 
 void resizeWindow(s32 w, s32 h, bool fullscreen);
 void onWindowResized(s32 w, s32 h);
